Force Rubocop version 0.54.0 in Hound CI
We use Hound CI to validate our commits against the style guide
contained in this repository. Unfortunately, Rubocop version bumps
often introduce breaking changes, which may cause programmers'
headaches.

For this reason, we are freezing Rubocop in Hound CI at version 0.54.0,
which is also a current default in Hound CI.